refactor : 헥사고날 제거
