+++ STORIX-Domain/src/main/java/com/storix/domain/domains/hashtag/service/HashtagRecommendService.java
@@ -1,7 +1,8 @@
 package com.storix.domain.domains.hashtag.service;
 
-import com.storix.domain.domains.hashtag.application.port.LoadHashtagPort;
+import com.storix.domain.domains.hashtag.adaptor.HashtagAdaptor;
 import com.storix.domain.domains.hashtag.dto.HashtagRecommendResponseDto;
+import com.storix.domain.domains.user.adaptor.UserAdaptor;
 import com.storix.domain.domains.user.application.port.LoadUserPort;
 import com.storix.domain.domains.user.domain.User;
 import com.storix.domain.domains.works.domain.Genre;
@@ -18,31 +19,32 @@ import java.util.Set;
 public class HashtagRecommendService {
 
     private final LoadUserPort loadUserPort;
-    private final LoadHashtagPort loadHashtagPort;
+    private final HashtagAdaptor hashtagAdaptor;
+    private final UserAdaptor userAdaptor;
 
     private static final int RECOMMENDATION_LIMIT = 10; // 추천 개수 제한
 
-    public List<HashtagRecommendResponseDto> getRecommendedHashtags(Long userId) {
-        // 1. 비로그인 유저(Guest) -> 전체 인기 태그 추천
-        if (userId == null) {
-            return loadHashtagPort.recommendGlobalPopular(RECOMMENDATION_LIMIT);
-        }
+    public List<HashtagRecommendResponseDto> getGlobalPopularHashtags() {
+        return hashtagAdaptor.recommendGlobalPopular(RECOMMENDATION_LIMIT);
+    }
 
-        // 2. 로그인 유저 정보 로드
-        User user = loadUserPort.findById(userId);
+    public List<HashtagRecommendResponseDto> getRecommendedHashtags(Long userId) {
+
+        // 1. 로그인 유저 정보 로드
+        User user = userAdaptor.findUserById(userId);
         Set<Genre> favoriteGenres = user.getFavoriteGenreList();
 
-        // 3. 선호 장르가 설정되지 않은 유저 -> 전체 인기 태그 추천
+        // 2. 선호 장르가 설정되지 않은 유저 -> 전체 인기 태그 추천
         if (favoriteGenres == null || favoriteGenres.isEmpty()) {
-            return loadHashtagPort.recommendGlobalPopular(RECOMMENDATION_LIMIT);
+            return this.getGlobalPopularHashtags();
         }
 
-        // 4. 선호 장르 기반 추천
+        // 3. 선호 장르 기반 추천
         List<HashtagRecommendResponseDto> recommendations =
-                loadHashtagPort.recommendByGenres(favoriteGenres, RECOMMENDATION_LIMIT);
+                hashtagAdaptor.recommendByGenres(favoriteGenres, RECOMMENDATION_LIMIT);
 
         if (recommendations.isEmpty()) {
-            return loadHashtagPort.recommendGlobalPopular(RECOMMENDATION_LIMIT);
+            return this.getGlobalPopularHashtags();
         }
 
         return recommendations;
